Move pending in queue notification
