+++ site/src/pages/WorkspacePage/Workspace.tsx
@@ -1,9 +1,8 @@
 import { type Interpolation, type Theme } from "@emotion/react";
 import Button from "@mui/material/Button";
 import AlertTitle from "@mui/material/AlertTitle";
-import { type FC, useEffect, useState } from "react";
+import { type FC, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import dayjs from "dayjs";
 import type * as TypesGen from "api/typesGenerated";
 import { Alert, AlertDetail } from "components/Alert/Alert";
 import { Stack } from "components/Stack/Stack";
@@ -85,43 +84,6 @@ export const Workspace: FC<WorkspaceProps> = ({
 }) => {
   const navigate = useNavigate();
   const theme = useTheme();
-
-  const [showAlertPendingInQueue, setShowAlertPendingInQueue] = useState(false);
-
-  // 2023-11-15 - MES - This effect will be called every single render because
-  // "now" will always change and invalidate the dependency array. Need to
-  // figure out if this effect really should run every render (possibly meaning
-  // no dependency array at all), or how to get the array stabilized (ideal)
-  const now = dayjs();
-  useEffect(() => {
-    if (
-      workspace.latest_build.status !== "pending" ||
-      workspace.latest_build.job.queue_size === 0
-    ) {
-      if (!showAlertPendingInQueue) {
-        return;
-      }
-
-      const hideTimer = setTimeout(() => {
-        setShowAlertPendingInQueue(false);
-      }, 250);
-      return () => {
-        clearTimeout(hideTimer);
-      };
-    }
-
-    const t = Math.max(
-      0,
-      5000 - dayjs().diff(dayjs(workspace.latest_build.created_at)),
-    );
-    const showTimer = setTimeout(() => {
-      setShowAlertPendingInQueue(true);
-    }, t);
-
-    return () => {
-      clearTimeout(showTimer);
-    };
-  }, [workspace, now, showAlertPendingInQueue]);
 
   const transitionStats =
     template !== undefined ? ActiveTransition(template, workspace) : undefined;
@@ -238,24 +200,6 @@ export const Workspace: FC<WorkspaceProps> = ({
               />
             )}
 
-            {showAlertPendingInQueue && (
-              <Alert severity="info">
-                <AlertTitle>Workspace build is pending</AlertTitle>
-                <AlertDetail>
-                  <div css={styles.alertPendingInQueue}>
-                    This workspace build job is waiting for a provisioner to
-                    become available. If you have been waiting for an extended
-                    period of time, please contact your administrator for
-                    assistance.
-                  </div>
-                  <div>
-                    Position in queue:{" "}
-                    <strong>{workspace.latest_build.job.queue_position}</strong>
-                  </div>
-                </AlertDetail>
-              </Alert>
-            )}
-
             {workspace.latest_build.job.error && (
               <Alert
                 severity="error"
@@ -357,9 +301,5 @@ const styles = {
 
   firstColumnSpacer: {
     flex: 2,
-  },
-
-  alertPendingInQueue: {
-    marginBottom: 12,
   },
 } satisfies Record<string, Interpolation<Theme>>;
