+++ site/src/pages/WorkspacePage/WorkspaceNotifications.tsx
@@ -3,10 +3,11 @@ import Button from "@mui/material/Button";
 import { workspaceResolveAutostart } from "api/queries/workspaceQuota";
 import { TemplateVersion, Workspace } from "api/typesGenerated";
 import { Alert, AlertDetail } from "components/Alert/Alert";
-import { FC } from "react";
+import { FC, useEffect, useState } from "react";
 import { useQuery } from "react-query";
 import { WorkspacePermissions } from "./permissions";
 import { DormantWorkspaceBanner } from "./DormantWorkspaceBanner";
+import dayjs from "dayjs";
 
 type WorkspaceNotificationsProps = {
   workspace: Workspace;
@@ -31,6 +32,43 @@ export const WorkspaceNotifications: FC<WorkspaceNotificationsProps> = (
     workspace.outdated;
   const autoStartFailing = workspace.autostart_schedule && !canAutostart;
   const requiresManualUpdate = updateRequired && autoStartFailing;
+
+  // Pending in Queue
+  const [showAlertPendingInQueue, setShowAlertPendingInQueue] = useState(false);
+  // 2023-11-15 - MES - This effect will be called every single render because
+  // "now" will always change and invalidate the dependency array. Need to
+  // figure out if this effect really should run every render (possibly meaning
+  // no dependency array at all), or how to get the array stabilized (ideal)
+  const now = dayjs();
+  useEffect(() => {
+    if (
+      workspace.latest_build.status !== "pending" ||
+      workspace.latest_build.job.queue_size === 0
+    ) {
+      if (!showAlertPendingInQueue) {
+        return;
+      }
+
+      const hideTimer = setTimeout(() => {
+        setShowAlertPendingInQueue(false);
+      }, 250);
+      return () => {
+        clearTimeout(hideTimer);
+      };
+    }
+
+    const t = Math.max(
+      0,
+      5000 - dayjs().diff(dayjs(workspace.latest_build.created_at)),
+    );
+    const showTimer = setTimeout(() => {
+      setShowAlertPendingInQueue(true);
+    }, t);
+
+    return () => {
+      clearTimeout(showTimer);
+    };
+  }, [workspace, now, showAlertPendingInQueue]);
 
   return (
     <>
@@ -81,6 +119,23 @@ export const WorkspaceNotifications: FC<WorkspaceNotificationsProps> = (
         )}
 
       <DormantWorkspaceBanner workspace={workspace} />
+
+      {showAlertPendingInQueue && (
+        <Alert severity="info">
+          <AlertTitle>Workspace build is pending</AlertTitle>
+          <AlertDetail>
+            <div css={{ marginBottom: 12 }}>
+              This workspace build job is waiting for a provisioner to become
+              available. If you have been waiting for an extended period of
+              time, please contact your administrator for assistance.
+            </div>
+            <div>
+              Position in queue:{" "}
+              <strong>{workspace.latest_build.job.queue_position}</strong>
+            </div>
+          </AlertDetail>
+        </Alert>
+      )}
     </>
   );
 };
